# marketplace/server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Comprehensive RESTful API for marketplace operations with real-time capabilities
  contact:
    name: Marketplace API Support
    email: support@marketplace.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.marketplace.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/auctions:
    get:
      tags: [Auctions]
      summary: List active auctions
      operationId: getAuctions
      security: []
      responses:
        '200':
          description: List of active auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      auctions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Auction'

    post:
      tags: [Auctions]
      summary: Create new auction
      operationId: createAuction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuctionRequest'
      responses:
        '201':
          description: Auction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      auction:
                        $ref: '#/components/schemas/Auction'
  /api/auctions/{id}/bid:
    post:
      tags: [Auctions]
      summary: Place bid on auction
      operationId: placeBid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBidRequest'
      responses:
        '200':
          description: Bid placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      bid:
                        $ref: '#/components/schemas/Bid'
                      auction:
                        $ref: '#/components/schemas/Auction'

  # Analytics Endpoints (Admin only)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Auction:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        title:
          type: string
        currentPrice:
          type: number
          format: float
        startPrice:
          type: number
          format: float
        reservePrice:
          type: number
          format: float
        bidIncrement:
          type: number
          format: float
        highestBidder:
          type: string
        totalBids:
          type: integer
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, active, ended, cancelled, sold]
        timeRemaining:
          type: integer
          description: Time remaining in milliseconds
        watcherCount:
          type: integer
    CreateAuctionRequest:
      type: object
      required: [itemId, startPrice, duration]
      properties:
        itemId:
          type: string
        startPrice:
          type: number
          format: float
          minimum: 0
        reservePrice:
          type: number
          format: float
          minimum: 0
        bidIncrement:
          type: number
          format: float
          minimum: 0.01
          default: 1.00
        duration:
          type: integer
          description: Duration in seconds
          minimum: 3600
          maximum: 2592000
        autoExtend:
          type: boolean
          default: false
        buyItNowPrice:
          type: number
          format: float
          minimum: 0
    Bid:
      type: object
      properties:
        id:
          type: string
        auctionId:
          type: string
        bidderId:
          type: string
        amount:
          type: number
          format: float
        maxBid:
          type: number
          format: float
        timestamp:
          type: string
          format: date-time
        isWinning:
          type: boolean
        isAutoBid:
          type: boolean
    PlaceBidRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: float
          minimum: 0
        maxBid:
          type: number
          format: float
          minimum: 0
...